ファイルをクラスで分割
